:bug: StudyRepository countBy 반환 값 Long으로 수정 및 그에 따른 service 코드 수정

+++ src/main/java/com/sejong/sejongpeer/domain/study/repository/StudyRepository.java
@@ -5,6 +5,7 @@ import com.sejong.sejongpeer.domain.study.entity.type.StudyType;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.time.LocalDateTime;
 
@@ -12,5 +13,5 @@ public interface StudyRepository extends JpaRepository<Study, Long>, StudyReposi
 
 	Slice<Study> findByTypeAndCreatedAtBetween(StudyType studyType, LocalDateTime startDate, LocalDateTime endDate, Pageable pageable);
 
-	int countByTypeAndCreatedAtBetween(StudyType studyType, LocalDateTime startDate, LocalDateTime endDate);
+	Long countByTypeAndCreatedAtBetween(StudyType studyType, LocalDateTime startDate, LocalDateTime endDate);
 }
